refactor: use black bg instead of transparent
Transparent bg introduces an alpha channel which speficifies the opacity for the pixels in an image. Albeit to normal users the bg is transparent, to the model during training, the background is perfectly visible as evident in the generated output of the model which discards the alpha channel and thus, revealing all pixels including the transparent ones.

This is also evident during data augmentation with Fiftyone as the augmented images do not include an alpha channel, resulting in visible backgrounds which may introduce unnecessary noise for the model.