feat: vehicle and lamp segmentation to determine make and model specific lamp positions
Vehicle will first be detected and segmented using a YOLOv8 model trained using the vehicle segmentation dataset. After which, its output will be passed as input to the lamp segmentation model to detect and segment lamp positions of a specifc make and model. The final output from the lamp segmentation is a cropped image of the original video frame, which is padded to match the original video resolution and written to an output video