feat: get mean pixel value to determine if lamp is lit up
Mean of all pixels within a cropped image of the lamp is obtained while excluding the absolute black pixels that are part of the binary mask. If mean > 200, the image is bright as the lamp is deemed as lit up.
However, one must take note that the threshold value used to determine if the lamp is lit up (in this case 200) depends on the type of lamp. In the case of the side lamp of a PRIUS 5DR HATCHBACK, even when the side lamp is turned on, the mean is only around 50~60 as only a small part of the side lamp is lit up instead of the entire strip. Hence, the threshold value is not universal.